Test (#99)
* Update preinstall.ps1

* Update preinstall.ps1

* mstarpkg

* Update config.yml

* testing

* removed cuda

* Inventor (#61)

* autodeskchgs

* added .net to preinstall

* Update preinstall.ps1

* Update preinstall.ps1

* kristrytofixmergeissue (#63)

* version update (#64)

* Update config.yml (#65)

* Update config.yml (#66)

* ansysr3pkg (#67)

* updated version (#68)

* first try google drive file stream (#69)

* version update (#70)

* Mstar (#71)

* Update config.yml

* updated lic file

* removed msmpisetup.exe blows up

* Update preinstall.ps1

* updated version

* Update config.yml

* Update config.yml

* Update config.yml

* reverted back to older cfd 

graphics issue

* Update config.yml

* Gdfs (#72)

* Update postinstall.ps1

* startupscript for GoogleDriveFS

* startup.ps1

to open google drive file stream

* Update config.yml

* Update config.yml

* Update config.yml

* Update postinstall.ps1

* Update config.yml (#73)

* Gdfs (#75)

* Update postinstall.ps1

* startupscript for GoogleDriveFS

* startup.ps1

to open google drive file stream

* Update config.yml

* Update config.yml

* Update config.yml

* Update postinstall.ps1

* testing

* Update config.yml

* testing

* Update config.yml

* Add files via upload

* chgtoschtask

* deletedstartupps1


* Paraviewfix (#76)

* added paraview

* paraviewicon

* Update config.yml

* replace icon (#77)

* switched to local license (#78)

* Mstar (#79)

* switched to local license

* fixed license


* Fusion version update (#80)

* updated version

* Update config.yml

* fixed parameters (#83)

* reverted back to cloud license (#81)

* Run certain steps if running on AppStream resource

* updatedforcmstuff (#85)

* Googledrivefilestream (#86)

* updatedgoogledrive

* 35 and 36

version 35 needs to be installed and then 36 is an upgrade

* Update postinstall.ps1

moved uninstall

* Update config.yml

35 not needed?

* Update postinstall.ps1

added stop process

* Update config.yml

removed added program version in file name

* Update postinstall.ps1

really disable googleupdate

* Update postinstall.ps1

removed the stop process

* Inventor (#88)

* licupdate

* Update config.yml

* Update config.yml

* Autocad Lic Fix (#89)

* updated lic file

* Update config.yml

* Update config.yml

* Update config.yml

* Ansyslic (#87)

* license config change

* x

* Update config.yml

* revised

* Update config.yml (#90)

* Update config.yml (#91)

* Update config.yml (#92)

changed capitalization on line 12 since it was not showing up in the appcatalog

* Rstudio (#93)

* upgraded to new format

* Update config.yml

* updated version (#94)

* Matlab (#95)

* updated version

* testing faster matlab launch

* version update (#97)

* version update (#96)

* TLS 1.2

* Update chocolateyinstall.ps1

* Update config.yml

+++ packages/_common/chocolateyinstall.ps1
@@ -398,9 +398,11 @@ Function Main($TOOLS_DIR, $INSTALL_DIR, $CONFIG) {
         
         $app_entry = "INSERT INTO Applications (Name, AbsolutePath, DisplayName, IconFilePath, LaunchParameters, WorkingDirectory) VALUES (@name, @path, @display, @icon, @params, @workdir)"
         $applications = ($CONFIG.Applications | Get-Member -MemberType NoteProperty).Name
+        write-output "$applications"
         
         foreach ($application in $applications) {
-            Write-Output "Creating App Catalog Entry for $($application.DisplayName)"
+            Write-Output "$application"
+            Write-Output "Creating App Catalog Entry for $CONFIG.Applications.$application.DisplayName"
             
             $app_icon_src = [io.path]::combine($TOOLS_DIR, 'icons', "$application.png")
             $app_icon = Join-Path "$APP_ICONS" "$application.png"
@@ -421,7 +423,15 @@ Function Main($TOOLS_DIR, $INSTALL_DIR, $CONFIG) {
             }
         }
     }
-
+    
+    Function PostInstall {
+    
+        # Run Postinstall PowerShell script
+        Write-Output "Running postinstall.ps1..."
+        Invoke-Expression $(Join-Path "$TOOLS_DIR" 'postinstall.ps1')
+        
+    }
+    
     # Only run config.yml sections if specified
     If ($CONFIG.Environment){EnvVars}
     If ($CONFIG.ChocoPackages){Choco}
@@ -430,16 +440,23 @@ Function Main($TOOLS_DIR, $INSTALL_DIR, $CONFIG) {
     If ($CONFIG.Files){Files}
     If ($CONFIG.Services){Services}
     If ($CONFIG.ScheduledTasks){SchedTask}
-    If ($CONFIG.Scripts){SessionScripts}
-    If ($CONFIG.Applications){AppCatalog}
+    
+    # Only run if package install is on an AppStream resource
+    If ($ENV:AppStream_Resource_Type){
+        write-output "AppStream Resource"
+        If ($CONFIG.Scripts){SessionScripts}
+        If ($CONFIG.Applications){AppCatalog}
+        
+        PostInstall        
 
-    # Run Postinstall PowerShell script
-    Write-Output "Running postinstall.ps1..."
-    Invoke-Expression $(Join-Path "$TOOLS_DIR" 'postinstall.ps1')
-
-    # Remove any files that might have been added to the Startup directory
-    Write-Output "Removing any startup files"
-    Remove-Item -Recurse -Force "$STARTUP"
+        # Remove any files that might have been added to the Startup directory
+        Write-Output "Removing any startup files"
+        Remove-Item -Recurse -Force "$STARTUP"
+    }
+    Else {
+        write-output "Not AppStream resource"
+        PostInstall
+    }    
 
     # Remove all installation files from disk
     if ((Test-Path $INSTALL_DIR) -and ($Mode.ToLower() -ne 't')) {
